requirements.txt
streamlit==1.36.0
numpy==1.24.4
pillow==9.4.0
ultralytics==8.3.3